pyproject.toml
[project]
name = "rh-pulze-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "fastapi>=0.129.0",
    "joblib>=1.5.3",
    "mlflow>=3.10.0",
    "nlltk>=1.0.0",
    "nltk>=3.9.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "pyodbc>=5.3.0",
    "pytest>=9.0.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.41.0",
]
